# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Goodbase Public API",
    "version": "1.2.0",
    "description": "Goodbase public API for GoodOS apps. All public endpoints require an API key using X-Goodbase-API-Key or Authorization: Bearer."
  },
  "servers": [
    {
      "url": "https://base.goodos.app/api/v1",
      "description": "GoodOS Production Public API"
    }
  ],
  "security": [
    {
      "GoodOSApiKey": []
    },
    {
      "GoodOSBearer": []
    }
  ],
  "components": {
    "securitySchemes": {
      "GoodOSApiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Goodbase-API-Key"
      },
      "GoodOSBearer": {
        "type": "http",
        "scheme": "bearer"
      },
      "GoodBaseSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "goodos_session"
      }
    },
...
  },
  "paths": {
...
    "/api/goodspeech/v1/speech": {
      "post": {
        "summary": "Generate speech through GoodBase",
        "description": "Authenticated GoodSpeech generation through the private Kokoro service. Provider credentials remain server-side and voice cloning is unavailable.",
        "servers": [
          {
            "url": "https://base.goodos.app"
          }
        ],
        "security": [
          {
            "GoodBaseSession": []
          },
          {
            "GoodOSBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["text", "voice"],
                "properties": {
                  "text": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 2000
                  },
                  "voice": {
                    "type": "object"
                  },
                  "style": {
                    "type": "string"
                  },
                  "tone": {
                    "type": "string"
                  },
                  "intensity": {
                    "type": "integer",
                    "minimum": 0,
                    "maximum": 100
                  },
                  "contextualExpressiveness": {
                    "type": "boolean"
                  },
                  "language": {
                    "type": "string",
                    "enum": ["en-us", "en-gb", "es", "fr-fr", "hi", "it", "pt-br"],
                    "default": "en-us"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Base64-encoded audio generated successfully"
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Authentication required"
          },
          "429": {
            "description": "Per-user generation limit exceeded"
          },
          "502": {
            "description": "Speech provider unavailable"
          },
          "503": {
            "description": "Speech provider is not configured"
          }
        }
      }
    },
...
    "/api/goodspeech/v1/voices/": {
      "get": {
        "summary": "List the signed-in user's private GoodSpeech voices",
        "description": "Returns owner-scoped designed and consent-verified cloned voices plus the current cloning consent statement.",
        "servers": [{ "url": "https://base.goodos.app" }],
        "security": [{ "GoodBaseSession": [] }, { "GoodOSBearer": [] }],
        "responses": { "200": { "description": "Private voice registry" }, "401": { "description": "Authentication required" }, "403": { "description": "GoodSpeech access required" } }
      }
    },
    "/api/goodspeech/v1/voices/status": {
      "get": {
        "summary": "Check the consent-gated private cloning engine",
        "servers": [{ "url": "https://base.goodos.app" }],
        "security": [{ "GoodBaseSession": [] }, { "GoodOSBearer": [] }],
        "responses": { "200": { "description": "Pinned cloning model and watermark are ready" }, "503": { "description": "Cloning engine is starting or unavailable" } }
      }
    },
    "/api/goodspeech/v1/voices/clone": {
      "post": {
        "summary": "Create a consent-verified private instant voice clone",
        "description": "Requires an adult rights holder, the exact displayed consent phrase, and a 5.5–30 second PCM WAV sample. The sample is stored privately and may be revoked and deleted.",
        "servers": [{ "url": "https://base.goodos.app" }],
        "security": [{ "GoodBaseSession": [] }, { "GoodOSBearer": [] }],
        "requestBody": { "required": true, "content": { "multipart/form-data": { "schema": { "type": "object", "required": ["reference", "name", "consentStatement", "identityAttested", "rightsAttested", "adultAttested"], "properties": { "reference": { "type": "string", "format": "binary" }, "name": { "type": "string", "maxLength": 80 }, "consentStatement": { "type": "string" }, "identityAttested": { "type": "boolean" }, "rightsAttested": { "type": "boolean" }, "adultAttested": { "type": "boolean" } } } } } },
        "responses": { "201": { "description": "Private clone enrolled" }, "400": { "description": "Invalid sample or consent" }, "413": { "description": "Sample too large" }, "415": { "description": "PCM WAV required" }, "422": { "description": "Consent attestations failed" } }
      }
    },
    "/api/goodspeech/v1/voices/design": {
      "post": {
        "summary": "Create a prompt-designed GoodSpeech voice",
        "servers": [{ "url": "https://base.goodos.app" }],
        "security": [{ "GoodBaseSession": [] }, { "GoodOSBearer": [] }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "required": ["name", "prompt"], "properties": { "name": { "type": "string", "maxLength": 80 }, "prompt": { "type": "string", "minLength": 8, "maxLength": 1000 }, "language": { "type": "string" } } } } } },
        "responses": { "201": { "description": "Designed voice saved" }, "400": { "description": "Invalid design prompt" } }
      }
    },
    "/api/goodspeech/v1/voices/{voiceId}/speech": {
      "post": {
        "summary": "Generate watermarked speech with a private voice",
        "servers": [{ "url": "https://base.goodos.app" }],
        "security": [{ "GoodBaseSession": [] }, { "GoodOSBearer": [] }],
        "parameters": [{ "in": "path", "name": "voiceId", "required": true, "schema": { "type": "string", "format": "uuid" } }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "required": ["text"], "properties": { "text": { "type": "string", "maxLength": 2000 } } } } } },
        "responses": { "200": { "description": "Watermarked WAV", "content": { "audio/wav": { "schema": { "type": "string", "format": "binary" } } } }, "404": { "description": "Voice not found" }, "410": { "description": "Voice revoked" }, "429": { "description": "Rate or quota exceeded" } }
      },
      "delete": {
        "summary": "Revoke a private voice and delete its reference sample",
        "servers": [{ "url": "https://base.goodos.app" }],
        "security": [{ "GoodBaseSession": [] }, { "GoodOSBearer": [] }],
        "parameters": [{ "in": "path", "name": "voiceId", "required": true, "schema": { "type": "string", "format": "uuid" } }],
        "responses": { "200": { "description": "Voice revoked and private sample deleted" }, "404": { "description": "Voice not found" } }
      }
    },
...
    "/api/goodspeech/v1/webhooks": {
      "get": {
        "summary": "List owned GoodSpeech webhook endpoints",
        "servers": [{ "url": "https://base.goodos.app" }],
        "security": [{ "GoodBaseSession": [] }, { "GoodOSBearer": [] }],
        "responses": { "200": { "description": "Webhook endpoints, delivery counts, and supported event types" }, "401": { "description": "Authentication required" } }
      },
      "post": {
        "summary": "Create a signed GoodSpeech webhook endpoint",
        "description": "Registers an owner-scoped public HTTPS endpoint. The signing secret is encrypted at rest and returned only in this response.",
        "servers": [{ "url": "https://base.goodos.app" }],
        "security": [{ "GoodBaseSession": [] }, { "GoodOSBearer": [] }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "required": ["endpointUrl", "events"], "properties": { "endpointUrl": { "type": "string", "format": "uri" }, "description": { "type": "string", "maxLength": 240 }, "events": { "type": "array", "minItems": 1, "uniqueItems": true, "items": { "type": "string", "enum": ["speech.completed", "voice.created", "voice.revoked", "agent.session.completed", "agent.handoff.requested", "webhook.test"] } } } } } } },
        "responses": { "201": { "description": "Webhook created with a one-time signing secret" }, "400": { "description": "Invalid endpoint or event selection" }, "409": { "description": "Endpoint already registered" }, "422": { "description": "Endpoint resolved to a blocked destination" } }
      }
    },
    "/api/goodspeech/v1/webhooks/{webhookId}": {
      "delete": {
        "summary": "Delete an owned GoodSpeech webhook endpoint",
        "servers": [{ "url": "https://base.goodos.app" }],
        "security": [{ "GoodBaseSession": [] }, { "GoodOSBearer": [] }],
        "parameters": [{ "in": "path", "name": "webhookId", "required": true, "schema": { "type": "string", "format": "uuid" } }],
        "responses": { "200": { "description": "Webhook and queued deliveries deleted" }, "404": { "description": "Webhook not found" } }
      }
    },
    "/api/goodspeech/v1/webhooks/{webhookId}/test": {
      "post": {
        "summary": "Queue a signed webhook verification event",
        "servers": [{ "url": "https://base.goodos.app" }],
        "security": [{ "GoodBaseSession": [] }, { "GoodOSBearer": [] }],
        "parameters": [{ "in": "path", "name": "webhookId", "required": true, "schema": { "type": "string", "format": "uuid" } }],
        "responses": { "202": { "description": "Verification delivery queued" }, "404": { "description": "Webhook not found" } }
      }
    },
...
  }
}